Remove everything that is not a proto_library
The core of this repository consists of 4 proto files that essentially
act as an API specification. The build logic should reflect that.

It's is unrealistic to support all possible ways of generating protos for
every language and every build system. Instead of supporting every
possible use-case explicitly, shift focus on the actual proto files and
make it as easy as possible to depend on them.

This commit removes all language-specific code, updates the few
remaining dependencies to their latest versions and adds bzlmod support.